requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.8.2
requests==2.31.0
python-multipart==0.0.6
cerebras-cloud-sdk
python-dotenv==1.0.1
